+++ crates/polyfuse/src/reply.rs
@@ -1,91 +1,36 @@
 use self::non_exhaustive::NonExhaustive;
-use crate::{
-    conn::Writer,
-    types::{FileAttr, FileLock, FsStatistics},
-    util::as_bytes,
-    write,
-};
-use polyfuse_kernel as kernel;
-use std::{fmt, io, time::Duration};
+use crate::types::{FileAttr, FileLock, FsStatistics};
+use std::{io, time::Duration};
 
-pub type Ok = ();
+/// The error values caused by the filesystem.
+pub trait Error {
+    /// Construct the error value from an I/O error.
+    fn from_io_error(io_error: io::Error) -> Self;
 
-#[derive(Debug)]
-pub struct Error(ErrorKind);
-
-#[derive(Debug)]
-enum ErrorKind {
-    Code(i32),
-    Fatal(io::Error),
-}
-
-impl Error {
-    pub(crate) fn code(&self) -> Option<i32> {
-        match self.0 {
-            ErrorKind::Code(code) => Some(code),
-            _ => None,
-        }
-    }
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("OpError").finish()
-    }
-}
-
-impl std::error::Error for Error {}
-
-impl crate::error::Error for Error {
-    fn from_io_error(io_error: io::Error) -> Self {
-        Self(ErrorKind::Fatal(io_error))
-    }
-
+    /// Construct the error value from an OS error code.
+    #[inline]
     fn from_code(code: i32) -> Self
     where
         Self: Sized,
     {
-        Self(ErrorKind::Code(code))
+        Self::from_io_error(io::Error::from_raw_os_error(code))
     }
 }
 
-pub type Result<T = Ok, E = Error> = std::result::Result<T, E>;
-
-pub struct ReplyEntry<'req> {
-    pub(crate) writer: &'req Writer,
-    pub(crate) header: &'req kernel::fuse_in_header,
-    pub(crate) arg: kernel::fuse_entry_out,
+/// Equivalent to `<E as Error>::from_code(code)`.
+#[inline]
+pub fn error_code<E>(code: i32) -> E
+where
+    E: Error,
+{
+    E::from_code(code)
 }
 
-impl ReplyEntry<'_> {
-    pub fn entry(mut self, attr: &FileAttr, opts: &EntryOptions) -> Result {
-        fill_attr(attr, &mut self.arg.attr);
-        self.arg.nodeid = opts.ino;
-        self.arg.generation = opts.generation;
+pub trait ReplyEntry {
+    type Ok;
+    type Error: Error;
 
-        if let Some(ttl) = opts.ttl_attr {
-            self.arg.attr_valid = ttl.as_secs();
-            self.arg.attr_valid_nsec = ttl.subsec_nanos();
-        } else {
-            self.arg.attr_valid = u64::MAX;
-            self.arg.attr_valid_nsec = u32::MAX;
-        }
-
-        if let Some(ttl) = opts.ttl_entry {
-            self.arg.entry_valid = ttl.as_secs();
-            self.arg.entry_valid_nsec = ttl.subsec_nanos();
-        } else {
-            self.arg.entry_valid = u64::MAX;
-            self.arg.entry_valid_nsec = u32::MAX;
-        }
-
-        write::send_reply(self.writer, self.header.unique, unsafe {
-            as_bytes(&self.arg)
-        })
-        .map_err(crate::error::io)?;
-
-        Ok(())
-    }
+    fn entry(self, attr: &FileAttr, opts: &EntryOptions) -> Result<Self::Ok, Self::Error>;
 }
 
 /// The option values for `ReplyEntry`.
@@ -140,95 +85,34 @@ impl Default for EntryOptions {
     }
 }
 
-pub struct ReplyAttr<'req> {
-    pub(crate) writer: &'req Writer,
-    pub(crate) header: &'req kernel::fuse_in_header,
-    pub(crate) arg: kernel::fuse_attr_out,
+pub trait ReplyAttr {
+    type Ok;
+    type Error: Error;
+
+    fn attr(self, attr: &FileAttr, ttl: Option<Duration>) -> Result<Self::Ok, Self::Error>;
 }
 
-impl ReplyAttr<'_> {
-    pub fn attr(mut self, attr: &FileAttr, ttl: Option<Duration>) -> Result {
-        fill_attr(attr, &mut self.arg.attr);
+pub trait ReplyOk {
+    type Ok;
+    type Error: Error;
 
-        if let Some(ttl) = ttl {
-            self.arg.attr_valid = ttl.as_secs();
-            self.arg.attr_valid_nsec = ttl.subsec_nanos();
-        } else {
-            self.arg.attr_valid = u64::MAX;
-            self.arg.attr_valid_nsec = u32::MAX;
-        }
-
-        write::send_reply(self.writer, self.header.unique, unsafe {
-            as_bytes(&self.arg)
-        })
-        .map_err(crate::error::io)?;
-
-        Ok(())
-    }
+    fn ok(self) -> Result<Self::Ok, Self::Error>;
 }
 
-pub struct ReplyOk<'req> {
-    pub(crate) writer: &'req Writer,
-    pub(crate) header: &'req kernel::fuse_in_header,
-}
+pub trait ReplyData {
+    type Ok;
+    type Error: Error;
 
-impl ReplyOk<'_> {
-    pub fn ok(self) -> Result {
-        write::send_reply(self.writer, self.header.unique, &[]).map_err(crate::error::io)?;
-        Ok(())
-    }
-}
-
-pub struct ReplyData<'req> {
-    pub(crate) writer: &'req Writer,
-    pub(crate) header: &'req kernel::fuse_in_header,
-}
-
-impl ReplyData<'_> {
-    pub fn data<T>(self, data: T) -> Result
+    fn data<T>(self, data: T) -> Result<Self::Ok, Self::Error>
     where
-        T: AsRef<[u8]>,
-    {
-        write::send_reply(self.writer, self.header.unique, data.as_ref())
-            .map_err(crate::error::io)?;
-
-        Ok(())
-    }
+        T: AsRef<[u8]>;
 }
 
-pub struct ReplyOpen<'req> {
-    pub(crate) writer: &'req Writer,
-    pub(crate) header: &'req kernel::fuse_in_header,
-    pub(crate) arg: kernel::fuse_open_out,
-}
+pub trait ReplyOpen {
+    type Ok;
+    type Error: Error;
 
-impl ReplyOpen<'_> {
-    pub fn open(mut self, fh: u64, opts: &OpenOptions) -> Result {
-        self.arg.fh = fh;
-
-        if opts.direct_io {
-            self.arg.open_flags |= kernel::FOPEN_DIRECT_IO;
-        }
-
-        if opts.keep_cache {
-            self.arg.open_flags |= kernel::FOPEN_KEEP_CACHE;
-        }
-
-        if opts.nonseekable {
-            self.arg.open_flags |= kernel::FOPEN_NONSEEKABLE;
-        }
-
-        if opts.cache_dir {
-            self.arg.open_flags |= kernel::FOPEN_CACHE_DIR;
-        }
-
-        write::send_reply(self.writer, self.header.unique, unsafe {
-            as_bytes(&self.arg)
-        })
-        .map_err(crate::error::io)?;
-
-        Ok(())
-    }
+    fn open(self, fh: u64, opts: &OpenOptions) -> Result<Self::Ok, Self::Error>;
 }
 
 /// The option values for `ReplyOpen`.
@@ -266,232 +150,63 @@ impl Default for OpenOptions {
     }
 }
 
-pub struct ReplyWrite<'req> {
-    pub(crate) writer: &'req Writer,
-    pub(crate) header: &'req kernel::fuse_in_header,
-    pub(crate) arg: kernel::fuse_write_out,
+pub trait ReplyWrite {
+    type Ok;
+    type Error: Error;
+
+    fn size(self, size: u32) -> Result<Self::Ok, Self::Error>;
 }
 
-impl ReplyWrite<'_> {
-    pub fn size(mut self, size: u32) -> Result {
-        self.arg.size = size;
+pub trait ReplyStatfs {
+    type Ok;
+    type Error: Error;
 
-        write::send_reply(self.writer, self.header.unique, unsafe {
-            as_bytes(&self.arg)
-        })
-        .map_err(crate::error::io)?;
-
-        Ok(())
-    }
+    fn stat(self, stat: &FsStatistics) -> Result<Self::Ok, Self::Error>;
 }
 
-pub struct ReplyStatfs<'req> {
-    pub(crate) writer: &'req Writer,
-    pub(crate) header: &'req kernel::fuse_in_header,
-    pub(crate) arg: kernel::fuse_statfs_out,
-}
+pub trait ReplyXattr {
+    type Ok;
+    type Error: Error;
 
-impl ReplyStatfs<'_> {
-    pub fn stat(mut self, stat: &FsStatistics) -> Result {
-        fill_statfs(stat, &mut self.arg.st);
+    fn size(self, size: u32) -> Result<Self::Ok, Self::Error>;
 
-        write::send_reply(self.writer, self.header.unique, unsafe {
-            as_bytes(&self.arg)
-        })
-        .map_err(crate::error::io)?;
-
-        Ok(())
-    }
-}
-
-pub struct ReplyXattr<'req> {
-    pub(crate) writer: &'req Writer,
-    pub(crate) header: &'req kernel::fuse_in_header,
-    pub(crate) arg: kernel::fuse_getxattr_out,
-}
-
-impl ReplyXattr<'_> {
-    pub fn size(mut self, size: u32) -> Result {
-        self.arg.size = size;
-
-        write::send_reply(self.writer, self.header.unique, unsafe {
-            as_bytes(&self.arg)
-        })
-        .map_err(crate::error::io)?;
-
-        Ok(())
-    }
-
-    pub fn data<T>(self, data: T) -> Result
+    fn data<T>(self, data: T) -> Result<Self::Ok, Self::Error>
     where
-        T: AsRef<[u8]>,
-    {
-        write::send_reply(self.writer, self.header.unique, data.as_ref())
-            .map_err(crate::error::io)?;
-
-        Ok(())
-    }
+        T: AsRef<[u8]>;
 }
 
-pub struct ReplyLk<'req> {
-    pub(crate) writer: &'req Writer,
-    pub(crate) header: &'req kernel::fuse_in_header,
-    pub(crate) arg: kernel::fuse_lk_out,
+pub trait ReplyLk {
+    type Ok;
+    type Error: Error;
+
+    fn lk(self, lk: &FileLock) -> Result<Self::Ok, Self::Error>;
 }
 
-impl ReplyLk<'_> {
-    pub fn lk(mut self, lk: &FileLock) -> Result {
-        fill_lk(lk, &mut self.arg.lk);
+pub trait ReplyCreate {
+    type Ok;
+    type Error: Error;
 
-        write::send_reply(self.writer, self.header.unique, unsafe {
-            as_bytes(&self.arg)
-        })
-        .map_err(crate::error::io)?;
-
-        Ok(())
-    }
-}
-
-pub struct ReplyCreate<'req> {
-    pub(crate) writer: &'req Writer,
-    pub(crate) header: &'req kernel::fuse_in_header,
-    pub(crate) arg: CreateArg,
-}
-
-#[derive(Default)]
-#[repr(C)]
-pub(crate) struct CreateArg {
-    entry_out: kernel::fuse_entry_out,
-    open_out: kernel::fuse_open_out,
-}
-
-impl ReplyCreate<'_> {
-    pub fn create(
-        mut self,
+    fn create(
+        self,
         fh: u64,
         attr: &FileAttr,
         entry_opts: &EntryOptions,
         open_opts: &OpenOptions,
-    ) -> Result {
-        fill_attr(attr, &mut self.arg.entry_out.attr);
-        self.arg.entry_out.nodeid = entry_opts.ino;
-        self.arg.entry_out.generation = entry_opts.generation;
-
-        if let Some(ttl) = entry_opts.ttl_attr {
-            self.arg.entry_out.attr_valid = ttl.as_secs();
-            self.arg.entry_out.attr_valid_nsec = ttl.subsec_nanos();
-        } else {
-            self.arg.entry_out.attr_valid = u64::MAX;
-            self.arg.entry_out.attr_valid_nsec = u32::MAX;
-        }
-
-        if let Some(ttl) = entry_opts.ttl_entry {
-            self.arg.entry_out.entry_valid = ttl.as_secs();
-            self.arg.entry_out.entry_valid_nsec = ttl.subsec_nanos();
-        } else {
-            self.arg.entry_out.entry_valid = u64::MAX;
-            self.arg.entry_out.entry_valid_nsec = u32::MAX;
-        }
-
-        self.arg.open_out.fh = fh;
-
-        if open_opts.direct_io {
-            self.arg.open_out.open_flags |= kernel::FOPEN_DIRECT_IO;
-        }
-
-        if open_opts.keep_cache {
-            self.arg.open_out.open_flags |= kernel::FOPEN_KEEP_CACHE;
-        }
-
-        if open_opts.nonseekable {
-            self.arg.open_out.open_flags |= kernel::FOPEN_NONSEEKABLE;
-        }
-
-        if open_opts.cache_dir {
-            self.arg.open_out.open_flags |= kernel::FOPEN_CACHE_DIR;
-        }
-
-        write::send_reply(self.writer, self.header.unique, unsafe {
-            as_bytes(&self.arg)
-        })
-        .map_err(crate::error::io)?;
-
-        Ok(())
-    }
+    ) -> Result<Self::Ok, Self::Error>;
 }
 
-pub struct ReplyBmap<'req> {
-    pub(crate) writer: &'req Writer,
-    pub(crate) header: &'req kernel::fuse_in_header,
-    pub(crate) arg: kernel::fuse_bmap_out,
+pub trait ReplyBmap {
+    type Ok;
+    type Error: Error;
+
+    fn block(self, block: u64) -> Result<Self::Ok, Self::Error>;
 }
 
-impl ReplyBmap<'_> {
-    pub fn block(mut self, block: u64) -> Result {
-        self.arg.block = block;
+pub trait ReplyPoll {
+    type Ok;
+    type Error: Error;
 
-        write::send_reply(self.writer, self.header.unique, unsafe {
-            as_bytes(&self.arg)
-        })
-        .map_err(crate::error::io)?;
-
-        Ok(())
-    }
-}
-
-pub struct ReplyPoll<'req> {
-    pub(crate) writer: &'req Writer,
-    pub(crate) header: &'req kernel::fuse_in_header,
-    pub(crate) arg: kernel::fuse_poll_out,
-}
-
-impl ReplyPoll<'_> {
-    pub fn revents(mut self, revents: u32) -> Result {
-        self.arg.revents = revents;
-
-        write::send_reply(self.writer, self.header.unique, unsafe {
-            as_bytes(&self.arg)
-        })
-        .map_err(crate::error::io)?;
-
-        Ok(())
-    }
-}
-
-fn fill_attr(src: &FileAttr, dst: &mut kernel::fuse_attr) {
-    dst.ino = src.ino;
-    dst.size = src.size;
-    dst.mode = src.mode;
-    dst.nlink = src.nlink;
-    dst.uid = src.uid;
-    dst.gid = src.gid;
-    dst.rdev = src.rdev;
-    dst.blksize = src.blksize;
-    dst.blocks = src.blocks;
-    dst.atime = src.atime.secs;
-    dst.atimensec = src.atime.nsecs;
-    dst.mtime = src.mtime.secs;
-    dst.mtimensec = src.mtime.nsecs;
-    dst.ctime = src.ctime.secs;
-    dst.ctimensec = src.ctime.nsecs;
-}
-
-fn fill_statfs(src: &FsStatistics, dst: &mut kernel::fuse_kstatfs) {
-    dst.bsize = src.bsize;
-    dst.frsize = src.frsize;
-    dst.blocks = src.blocks;
-    dst.bfree = src.bfree;
-    dst.bavail = src.bavail;
-    dst.files = src.files;
-    dst.ffree = src.ffree;
-    dst.namelen = src.namelen;
-}
-
-fn fill_lk(src: &FileLock, dst: &mut kernel::fuse_file_lock) {
-    dst.typ = src.typ;
-    dst.start = src.start;
-    dst.end = src.end;
-    dst.pid = src.pid;
+    fn revents(self, revents: u32) -> Result<Self::Ok, Self::Error>;
 }
 
 mod non_exhaustive {
